requirements.txt
Flask==1.0.2
python-firebase==1.2
gunicorn==19.9.0
line-bot-sdk==1.8.0
beautifulsoup4==4.6.3
lxml==4.0.0
